+++ implementations/web/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @cotal-ai/web
 
+## 0.14.10
+
 ## 0.14.9
 
 ### Patch Changes